+++ packages/frontend/src/features/explorer/store/index.ts
@@ -15,4 +15,3 @@ export { explorerActions } from './explorerSlice';
 export * from './hooks';
 export * from './selectors';
 export { useExplorerInitialization } from './useExplorerInitialization';
-export * from './utils';